Re-org release scripts

+++ scripts/build_win_release.ps1
@@ -1,30 +1,29 @@
 
-$branch = "main"
-
-$working_dir = ".\tmp"
-$repo_dir = "$working_dir\repo"
-$artifact_dir = "$working_dir\artifacts"
-
-
-
-Remove-Item $working_dir -Recurse -Force
-New-Item -Path $working_dir -ItemType Directory
-New-Item -Path $repo_dir -ItemType Directory
-New-Item -Path $artifact_dir -ItemType Directory
+# $branch = "main"
+# $repo_url = "https://github.com/EtalumaSupport/LumaViewPro.git"
+$repo_url = "https://github.com/jmcoreymv/LumaViewPro-Private.git"
+$branch = "feature/add-windows-exe-gen"
 
 $version = Read-Host -Prompt "Set version"
 Write-Host "Version: $version"
 $lvp_base_w_version = "LumaViewPro-$version"
 
-echo "Cloning $branch for release"
-git clone --depth 1 --branch $branch https://github.com/EtalumaSupport/LumaViewPro.git $repo_dir
-Set-Location -Path $repo_dir
+$starting_dir = Get-Location
+$working_dir = ".\tmp"
+$repo_dir = "$working_dir\$lvp_base_w_version"
+$artifact_dir = "$working_dir\artifacts"
 
-# echo "Removing old artifacts if present..."
-# Remove-Item '.\dist' -Recurse -Force
-# Remove-Item '.\build' -Recurse -Force
-# Remove-Item '.\release' -Recurse -Force
-# Remove-Item '.\LICENSE.*' -Force
+if (Test-Path $working_dir) {
+    Remove-Item $working_dir -Recurse -Force
+}
+
+New-Item -Path $working_dir -ItemType Directory 
+New-Item -Path $repo_dir -ItemType Directory
+New-Item -Path $artifact_dir -ItemType Directory
+
+echo "Cloning $branch for release"
+git clone --depth 1 --branch $branch $repo_url $repo_dir
+Set-Location -Path $repo_dir
 
 echo "Creating release directory"
 New-Item -Path '.\release' -ItemType Directory
@@ -40,22 +39,23 @@ $compress = @{
 }
 Compress-Archive @compress
 
-echo "Creating .tar.gz bundle of source..."
-tar -czf ".\" "..\..\$artifact_dir\$lvp_base_w_version-source.tar.gz"
+# echo "Creating .tar.gz bundle of source..."
+# tar -czf ".\" "..\..\$artifact_dir\$lvp_base_w_version-source.tar.gz"
 
 echo "Generating .exe..."
+Copy-Item '.\scripts\config\lumaviewpro_win_release.spec' '.\lumaviewpro.spec'
 pyinstaller .\lumaviewpro.spec
 
 echo "Rename output directory"
 $orig_output_dir = ".\dist\lumaviewpro"
-$new_output_dir = ".\dist\LumaViewPro-$version"
-Rename-Item -Path $orig_output_dir -NewName $new_output_dir
+Rename-Item -Path $orig_output_dir -NewName $lvp_base_w_version
 
 echo "Creating .zip bundle of executable..."
 $compress = @{
-    Path = $new_output_dir
+    Path = ".\dist\$lvp_base_w_version"
     CompressionLevel = "Optimal"
     DestinationPath =  "..\..\$artifact_dir\$lvp_base_w_version.zip"
 }
 Compress-Archive @compress
 
+Set-Location $starting_dir
